Huge overhaul
refactored all routes and protected games routes
